meson: Produce versioned symbols for -Dlegacy=enabled
There's a discrepancy when doing a legacy enabled build with
autoconf and meson. For the autoconf case, resulting libxml2.so
contains versioned symbols. But that's not the case with meson
build. Pass necessary linker arguments from meson too.

Now, while Windows does support versioned symbols the symbols
file needs to be in a slightly different format. This is usually
achieved by preprocessing .syms file and producing new .def file
in expected format. Well, leave that for future work.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>