python: Fix bug #889 by only requesting len/4 bytes/characters
This ensures that if a string is returned, the buffer is large enough
to contain the UTF-8 encoded bytes.

The downside is that this doesn't utilize the full buffer space in
the common case where read() returns bytes or when most of the text
is ASCII which can be encoded with 1 byte per character. Potentially,
this makes parsing slower, though I haven't benchmarked it.